BREAAKING Stop using SCSS because the whole @import thing broke
I just used the sass program to convert everything into CSS. I do not
change the styles frequently, so it should not be too bad to maintain
the plan CSS going forward.

The output still looks messy, but I will clean it up later. The
important work is to keep the website the way it looked.